[Copilot] Extract lifecycle entities to separate module
