+++ src/vivarium/framework/lifecycle/__init__.py
@@ -7,15 +7,17 @@ This package provides tools for managing the simulation lifecycle.
 
 """
 
+from vivarium.framework.lifecycle.entities import (
+    LifeCycle,
+    LifeCyclePhase,
+    LifeCycleState,
+)
 from vivarium.framework.lifecycle.exceptions import (
     ConstraintError,
     InvalidTransitionError,
     LifeCycleError,
 )
 from vivarium.framework.lifecycle.lifecycle import (
-    LifeCycle,
     LifeCycleInterface,
     LifeCycleManager,
-    LifeCyclePhase,
-    LifeCycleState,
 )
